added initial installation instructions
This marks the beginning to the installation instructions in the README file. Downloading Godot engine is fairly straight forward - it is offered on multiple operating systems, with links for the most popular. The Initial 'Fishing Game - Godot' folder is in the process of review prior to being added to the main branch. The instructions can be built upon if more issues arise when trying to access the game files.